FIX: config.sh problems

+++ config.sh
@@ -34,7 +34,7 @@ fi
 
 echo "Installing mch packages &" $hostname "config files"
 
-if [[ -n ${reposdir} ]]; then
+if [[ -n ${reposdir} ]] && [[ ! -f "${reposdir}/repos.yaml" ]]; then
     echo " -$PWD" >> repos.yaml
     cp repos.yaml $reposdir/
 fi
